// src/domain/enums.ts
export enum Role {
  SUPERUSER = 'SUPERUSER',
  TECHNICIAN = 'TECHNICIAN',
  CLIENT = 'CLIENT',
}

export enum TicketStatus {
  OPEN = 'OPEN',
  IN_PROGRESS = 'IN_PROGRESS',
  RESOLVED = 'RESOLVED',
  CANCELLED = 'CANCELLED',
}

export enum NotificationType {
  TICKET_ASSUMED = 'TICKET_ASSUMED',
  TICKET_RETURNED = 'TICKET_RETURNED',
  TICKET_RESOLVED = 'TICKET_RESOLVED',
  TICKET_CANCELLED = 'TICKET_CANCELLED',
  TECHNICIAN_DEACTIVATED = 'TECHNICIAN_DEACTIVATED',
  COMMENT_CREATED = 'COMMENT_CREATED',
}

export enum UserStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
}